+++ packages/jdm-editor/package.json
@@ -34,6 +34,7 @@
     "build:watch": "vite build --watch",
     "storybook": "storybook dev -p 9009",
     "build:storybook": "storybook build -o docs",
+    "typecheck": "tsc --noEmit",
     "prepublishOnly": "vite build"
   },
   "dependencies": {
