+++ packages/jdm-editor/src/components/code-editor/local-ce.tsx
@@ -1,11 +1,12 @@
 import React, { useEffect, useState } from 'react';
 
 import { useDecisionGraphRaw } from '../decision-graph';
+import type { CodeEditorRef } from './ce';
 import { CodeEditor, type CodeEditorProps } from './ce';
 
 type LocalCodeEditorProps = Omit<CodeEditorProps, 'extension'>;
 
-export const LocalCodeEditor = React.forwardRef<HTMLDivElement, LocalCodeEditorProps>((props, ref) => {
+export const LocalCodeEditor = React.forwardRef<CodeEditorRef, LocalCodeEditorProps>((props, ref) => {
   const raw = useDecisionGraphRaw();
   const [extension, setExtension] = useState<CodeEditorProps['extension']>(() => {
     if (raw?.listenerStore) {
