+++ packages/jdm-editor/src/components/decision-graph/nodes/specifications/expression.specification.tsx
@@ -13,8 +13,8 @@ import type { expressionNodeSchema } from '../../../../helpers/schema';
 import { DiffInput, DiffRadio, DiffSwitch } from '../../../shared';
 import { DiffCodeEditor } from '../../../shared/diff-ce';
 import { useDecisionGraphActions, useDecisionGraphState, useNodeDiff } from '../../context/dg-store.context';
-import { compareAndUnifyLists, compareStringFields } from '../../dg-diff-util';
 import { type Diff, type DiffMetadata } from '../../dg-types';
+import { compareAndUnifyLists, compareStringFields } from '../../diff/comparison';
 import { TabExpression } from '../../graph/tab-expression';
 import { GraphNode } from '../graph-node';
 import type { NodeDecisionTableData } from './decision-table.specification';
