+++ packages/jdm-editor/src/components/decision-graph/nodes/specifications/specifications.tsx
@@ -3,8 +3,8 @@ import { expressionSpecification } from './expression.specification';
 import { functionSpecification } from './function.specification';
 import { inputSpecification } from './input.specification';
 import { outputSpecification } from './output.specification';
-import type { NodeSpecification } from './specification-types';
 import { NodeKind } from './specification-types';
+import type { NodeSpecification } from './specification-types';
 import { switchSpecification } from './switch.specification';
 
 function makeNodeSpecification<T extends Record<NodeKind, V>, V extends NodeSpecification>(o: T): Readonly<T> {
