+++ packages/jdm-editor/src/components/decision-graph/nodes/specifications/switch.specification.tsx
@@ -13,8 +13,8 @@ import { P, match } from 'ts-pattern';
 import { useNodeType } from '../../../../helpers/node-type';
 import { DiffCodeEditor } from '../../../shared/diff-ce';
 import { useDecisionGraphActions, useDecisionGraphState } from '../../context/dg-store.context';
-import { compareAndUnifyLists } from '../../dg-diff-util';
 import type { Diff, DiffMetadata } from '../../dg-types';
+import { compareAndUnifyLists } from '../../diff/comparison';
 import type { SimulationTrace, SimulationTraceDataSwitch } from '../../simulator/simulation.types';
 import { GraphNode } from '../graph-node';
 import { NodeColor } from './colors';
