+++ packages/jdm-editor/src/components/decision-table/dt.stories.tsx
@@ -128,7 +128,7 @@ const meta: Meta<typeof DecisionTable> = {
   argTypes: {
     configurable: { control: 'boolean' },
     disabled: { control: 'boolean' },
-    inputDataType: { control: 'object' },
+    inputData: { control: 'object' },
     cellRenderer: {
       control: false,
     },
@@ -137,7 +137,7 @@ const meta: Meta<typeof DecisionTable> = {
     inputsSchema: inputSchemaDefault,
     configurable: true,
     disabled: false,
-    inputDataType: {
+    inputData: {
       cart: { weight: 100, total: 100 },
       customer: { country: 'US', firstName: 'John', lastName: 'Doe' },
     },
